Move self sign helper and CCM under pkg/
They are not parts of DTLS, but useful enough to be exposed.
